Move created files to a less intusive place
