+++ .gitignore
@@ -4,7 +4,7 @@
 /db/admins.json
 /db/classes.db
 /db/tasks.json
-/security/certificates/
+/server/certificates/*
 
 # vim
 *.swp
